HDDS-15493. Limit checkpoint-format parameterization in TestOMRatisSnapshots to transfer tests
